requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
djangorestframework_simplejwt==5.2.2
djangorestframework-jwt==1.11.0
django-environ==0.10.0
channels==4.0.0
channels_redis==4.0.0
daphne==4.0.0